requirements.txt
numpy~=1.23.1
pandas~=1.5.2
PyQt5
paho-mqtt~=1.6.1
matplotlib~=3.6.2
xlwt
openpyxl
requests~=2.27.1
PyQt6~=6.5.1
sklearn~=0.0.post1
scikit-learn~=1.2.0
schedule~=1.2.0
PyGithub~=1.58.1